requirements.txt
django == 1.10
mimeparse == 0.1.3
python-dateutil == 2.5.3
six == 1.10.0
